// szamlazzponthu-test-ui-app/src/app/core/services/api/models/UsrCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Usr API
 * CRUD műveletek a Usr erőforráson.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UsrCreate
 */
export interface UsrCreate {
    /**
     * 
     * @type {string}
     * @memberof UsrCreate
     */
    firstname: string;
    /**
     * 
     * @type {string}
     * @memberof UsrCreate
     */
    lastname: string;
    /**
     * 
     * @type {string}
     * @memberof UsrCreate
     */
    address: string;
    /**
     * 
     * @type {string}
     * @memberof UsrCreate
     */
    telephone: string;
    /**
     * 
     * @type {boolean}
     * @memberof UsrCreate
     */
    active: boolean;
    /**
     * 
     * @type {string}
     * @memberof UsrCreate
     */
    job: string;
}

/**
 * Check if a given object implements the UsrCreate interface.
 */
export function instanceOfUsrCreate(value: object): value is UsrCreate {
    if (!('firstname' in value) || value['firstname'] === undefined) return false;
    if (!('lastname' in value) || value['lastname'] === undefined) return false;
    if (!('address' in value) || value['address'] === undefined) return false;
    if (!('telephone' in value) || value['telephone'] === undefined) return false;
    if (!('active' in value) || value['active'] === undefined) return false;
    if (!('job' in value) || value['job'] === undefined) return false;
    return true;
}

export function UsrCreateFromJSON(json: any): UsrCreate {
    return UsrCreateFromJSONTyped(json, false);
}

export function UsrCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): UsrCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'firstname': json['firstname'],
        'lastname': json['lastname'],
        'address': json['address'],
        'telephone': json['telephone'],
        'active': json['active'],
        'job': json['job'],
    };
}

export function UsrCreateToJSON(json: any): UsrCreate {
    return UsrCreateToJSONTyped(json, false);
}

export function UsrCreateToJSONTyped(value?: UsrCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'firstname': value['firstname'],
        'lastname': value['lastname'],
        'address': value['address'],
        'telephone': value['telephone'],
        'active': value['active'],
        'job': value['job'],
    };
}