// szamlazzponthu-test-ui-app/src/app/core/services/api/models/UsrPut.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Usr API
 * CRUD műveletek a Usr erőforráson.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UsrPut
 */
export interface UsrPut {
    /**
     * 
     * @type {string}
     * @memberof UsrPut
     */
    firstname: string;
    /**
     * 
     * @type {string}
     * @memberof UsrPut
     */
    lastname: string;
    /**
     * 
     * @type {string}
     * @memberof UsrPut
     */
    address: string;
    /**
     * 
     * @type {string}
     * @memberof UsrPut
     */
    telephone: string;
    /**
     * 
     * @type {boolean}
     * @memberof UsrPut
     */
    active: boolean;
    /**
     * 
     * @type {string}
     * @memberof UsrPut
     */
    job: string;
}

/**
 * Check if a given object implements the UsrPut interface.
 */
export function instanceOfUsrPut(value: object): value is UsrPut {
    if (!('firstname' in value) || value['firstname'] === undefined) return false;
    if (!('lastname' in value) || value['lastname'] === undefined) return false;
    if (!('address' in value) || value['address'] === undefined) return false;
    if (!('telephone' in value) || value['telephone'] === undefined) return false;
    if (!('active' in value) || value['active'] === undefined) return false;
    if (!('job' in value) || value['job'] === undefined) return false;
    return true;
}

export function UsrPutFromJSON(json: any): UsrPut {
    return UsrPutFromJSONTyped(json, false);
}

export function UsrPutFromJSONTyped(json: any, ignoreDiscriminator: boolean): UsrPut {
    if (json == null) {
        return json;
    }
    return {
        
        'firstname': json['firstname'],
        'lastname': json['lastname'],
        'address': json['address'],
        'telephone': json['telephone'],
        'active': json['active'],
        'job': json['job'],
    };
}

export function UsrPutToJSON(json: any): UsrPut {
    return UsrPutToJSONTyped(json, false);
}

export function UsrPutToJSONTyped(value?: UsrPut | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'firstname': value['firstname'],
        'lastname': value['lastname'],
        'address': value['address'],
        'telephone': value['telephone'],
        'active': value['active'],
        'job': value['job'],
    };
}